More clang-tidy clean-ups (#431)
* More clang-tidy clean-ups

* Restore FileProvider::get_file_name & get_file_size

+++ node/silkworm/types/receipt_cbor.cpp
@@ -28,7 +28,7 @@ Bytes cbor_encode(const std::vector<Receipt>& v) {
     if (v.empty()) {
         encoder.write_null();
     } else {
-        encoder.write_array(v.size());
+        encoder.write_array(static_cast<int>(v.size()));
     }
 
     for (const Receipt& r : v) {
